requirements.txt
numpy
scipy
tensorflow

pyyaml

PySoundFile
python_speech_features
editdistance
